refactor(utils/ , domains/): move dock utils in good place and rename file db connection
