+++ backend/src/domains/adminRequest/adminRequest.repository.js
@@ -1,6 +1,6 @@
 const pool = require('../../config/database');
 const Admin_requestMapper = require('./adminRequest.mapper');
-const utilsMapper = require('../utils/mapperUtils');
+const utilsMapper = require('../../utils/mapperUtils');
 
 class AdminRequestRepository {
     /**
